+++ benefits/RELEASES.md
@@ -8,3 +8,14 @@ Based on:
 - [typescript v1.0.0] benefits
 ### Releases
 - [NPM v1.0.0] https://www.npmjs.com/package/@wingspan/benefits/v/1.0.0 - benefits
+
+
+## 2023-11-13 01:25:15
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.0] benefits
+### Releases
+- [NPM v2.0.0] https://www.npmjs.com/package/@wingspan/benefits/v/2.0.0 - benefits